Create a VPM package version for Unity 2022
- Restructured the entire project to align with VRChat Package Manager (VPM) standards.
- Moved all files from /Assets/ to the new package format under /Packages/.
- Added package.json and other necessary package metadata files.
- Refactored file paths and references to support the new structure.
- Removed old prefabs and created new prefabs with new structure.
- Created scripts to facilitate installation of new prefabs into avatars.

This commit marks the transition from the old Unity asset-based installation to a fully-fledged VPM package, simplifying the installation and update process for users.

Note: We will continue to support the old asset-based style for those who prefer it, but we highly recommend transitioning to the new VPM package approach for a more streamlined and efficient experience.

We will also release a Unity 2019 VPM version **soon**!